move test_van to test_ngpvan
