+++ test/test_ngpvan/test_contact_notes.py
@@ -4,7 +4,7 @@ import unittest
 import requests_mock
 
 from parsons import VAN
-from test.test_van.responses_people import get_person_response
+from test.test_ngpvan.responses_people import get_person_response
 from test.utils import assert_matching_tables
 
 os.environ["VAN_API_KEY"] = "SOME_KEY"
